fix: One component per file for Tile (#441)
* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Changes to refactor Tile into separate source files

* Remove script file

* Fixed issues in playground

* Fixed issues in Panel docs page
