+++ src/Tile/_TileActions.js
@@ -1,0 +1,23 @@
+import classnames from 'classnames';
+import PropTypes from 'prop-types';
+import React from 'react';
+
+const TileActions = props => {
+    const { children, className, ...rest } = props;
+
+    const tileActionsClasses = classnames(
+        'fd-tile__actions',
+        className
+    );
+
+    return <div {...rest} className={tileActionsClasses}>{children}</div>;
+};
+
+TileActions.displayName = 'TileActions';
+
+TileActions.propTypes = {
+    children: PropTypes.node,
+    className: PropTypes.string
+};
+
+export default TileActions;
