+++ src/index.js
@@ -45,20 +45,15 @@ export { PanelFilters } from './Panel/Panel';
 export { PanelContent } from './Panel/Panel';
 export { PanelFooter } from './Panel/Panel';
 export { Popover } from './Popover/Popover';
-export { ProductTile } from './Tile/Tile';
-export { ProductTileContent } from './Tile/Tile';
-export { ProductTileMedia } from './Tile/Tile';
+export { default as ProductTile } from './Tile/ProductTile';
 export { SearchInput } from './SearchInput/SearchInput';
 export { SideNav, SideNavList, SideNavListItem } from './SideNavigation/SideNavigation';
 export { Tab } from './Tabs/Tab';
 export { TabGroup } from './Tabs/TabGroup';
 export { Table } from './Table/Table';
 export { Token } from './Token/Token';
-export { Tile } from './Tile/Tile';
-export { TileContent } from './Tile/Tile';
-export { TileMedia } from './Tile/Tile';
-export { TileActions } from './Tile/Tile';
-export { TileGrid } from './Tile/Tile';
+export { default as Tile } from './Tile/Tile';
+export { default as TileGrid } from './Tile/TileGrid';
 export { Toggle } from './Toggle/Toggle';
 export { default as TreeView } from './TreeView/TreeView';
 export { Time } from './Time/Time';
